+++ examples/data-sources/iosxe_crypto_ikev2_policy/data-source.tf
@@ -1,0 +1,3 @@
+data "iosxe_crypto_ikev2_policy" "example" {
+  name = "policy1"
+}