+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Add `iosxe_crypto_ipsec_profile` resource and data source
 - Add `iosxe_radius_server` resource and data source
 - Add `iosxe_crypto_ikev2_keyring` resource and data source
+- Add `iosxe_crypto_ikev2_policy` resource and data source
 
 ## 0.3.2